+++ xr3/h/XR/Renderer.hpp
@@ -7,7 +7,6 @@
 // copyright (c) Nuclear Heart Interactive Ltd. All rights reserved.
 //
 //==============================================================================
-
 #include "Rect.hpp"
 #include "Material.hpp"
 #include "FloatBuffer.hpp"
@@ -32,26 +31,26 @@ public:
     BF_COLOR = 0x01,
     BF_DEPTH = 0x02
   };
-  
+
   enum
   {
     kNumPersMatrixElems = 16
   };
-  
+
   // static
   static void           Init(void* mainWindow);
   static void           Exit();
-  
+
   static int32_t        GetScreenWidth();
   static int32_t        GetScreenHeight();
-  
+
   static int32_t        GetDeviceWidth();
   static int32_t        GetDeviceHeight();
-  
+
   static void*          Alloc(size_t bytes);
   static Material*      AllocMaterial();
-  static FloatBuffer*   AllocBuffer(size_t elemSize, size_t numElems);
-  
+  static FloatBuffer*   AllocBuffer(uint32_t elemSize, uint32_t numElems);
+
   static void           SetPerspMatrix(const float (&arData)[kNumPersMatrixElems]);
 
   ///@brief Sets up orthographic projection.
@@ -80,7 +79,7 @@ public:
 
   static void           SetViewMatrix(const Matrix& m);
   static void           SetModelMatrix(const Matrix& m);
-  
+
   static void           GetViewMatrix(Matrix& m);
   static void           GetModelMatrix(Matrix& m);
 
@@ -91,14 +90,14 @@ public:
 
   static void           DrawPrims(PrimType pt);
   static void           DrawPrims(PrimType pt, const uint16_t* pInds, int numInds);
-                  
+
   static void           PrintString(int16_t x, int16_t y, const char* pString);
   static void           SetClearColor(Color c);
   static void           SetAmbientColor(Color c);
-  
+
   static Color          GetClearColor();
   static Color          GetAmbientColor();
-  
+
   static Rect           GetScissorRect();
   static void           SetScissorRect( const Rect& r );
   static void           ClearScissorRect();
@@ -108,7 +107,7 @@ public:
   static void           SetFog(bool state);
   static void           SetFogColor(Color c);
   static void           SetFogRange(float zFar, float zNear);
-  
+
   static void           ClearBuffer(uint32_t flags = (BF_COLOR | BF_DEPTH));
   static void           Flush();
   static void           Present();
